VORTEX-5641 : Add/Remove Custom Columns (#29)
* Add ability to remove dynamic meta-data information.
Add ability to modify dynamic user-created columns in List Tool table model.
Fix potential zip bug in About panel.
Fix numeric casting bug in Mean/Median functions.

* Fixed AboutPanel zip for Windows.

* Finalize column removal.

* Fix warnings.

* Add missing override method.

* VORTEX-5641 : All custom column cells editable.

* Resolve comments.

+++ open-sphere-base/mantle/src/main/java/io/opensphere/mantle/data/dynmeta/impl/DynamicMetadataStringController.java
@@ -33,14 +33,14 @@ public class DynamicMetadataStringController extends AbstractDynamicMetadataCont
         {
             for (Long id : cacheIds)
             {
-                String oldStringValue = getIdToValueMap().get(id);
+                String oldStringValue = getIdToValueMap().get(id.longValue());
                 if (oldStringValue == null)
                 {
-                    getIdToValueMap().put(id, stringValueToAppend);
+                    getIdToValueMap().put(id.longValue(), stringValueToAppend);
                 }
                 else
                 {
-                    getIdToValueMap().put(id, oldStringValue + stringValueToAppend);
+                    getIdToValueMap().put(id.longValue(), oldStringValue + stringValueToAppend);
                 }
             }
             fireChangeEvent(cacheIds, source);
